Fix bug related to sameLine, due to discrependancy between widgets size in a same sameLine call streak.
To fix bug it was necessary to properly separate what is done at the beginning of a widget with computeRelativePosition and at the end with updateSpacing.
Remove redundant computeSpacing and merge it with updateSpacing.
Change order in which description and inputColor/Vector are drawn.

There is a bug in beginWindow : ItemState is set to active for the active box, and not for the button to close/reduce it, so they are not functionnal. Just changing the order in which it is done give shitty jaggy movements.